requirements.txt
pandas==0.20.3
Numpy==1.12.1
facebook_business==3.1.2
googleads==14.0.0
pyyaml==3.12